+++ src/erasure-code/clay/ErasureCodeClay.h
@@ -72,7 +72,7 @@ public:
 
   unsigned int get_chunk_size(unsigned int stripe_width) const override;
 
-  unsigned int get_minimum_granularity() override;
+  size_t get_minimum_granularity() override;
 
   int minimum_to_decode(const std::set<int> &want_to_read,
 			const std::set<int> &available,
