+++ src/erasure-code/lrc/ErasureCodeLrc.cc
@@ -559,7 +559,7 @@ unsigned int ErasureCodeLrc::get_chunk_size(unsigned int stripe_width) const
   return layers.front().erasure_code->get_chunk_size(stripe_width);
 }
 
-unsigned int ErasureCodeLrc::get_minimum_granularity()
+size_t ErasureCodeLrc::get_minimum_granularity()
 {
   return layers.front().erasure_code->get_minimum_granularity();
 }
