+++ src/erasure-code/shec/ErasureCodeShec.h
@@ -153,7 +153,7 @@ public:
                    std::map <int, ceph::bufferptr> &out);
 
   unsigned get_alignment() const override;
-  unsigned int get_minimum_granularity() override
+  size_t get_minimum_granularity() override
   {
     return 1;
   }
